Simplify directory structure after switch to workspaces
