--- projects/rclio/src/rclio.rs
@@ -1,3 +1,0 @@
-mod all;
-
-pub use all::*;
