+++ projects/rooster/src/lib.rs
@@ -4,8 +4,8 @@ use crate::password::v2::PasswordStore;
 use clap::{App, AppSettings, Arg};
 use rclio::CliInputOutput;
 use rclio::OutputType;
-use rutil::rutil::safe_string::SafeString;
-use rutil::rutil::safe_vec::SafeVec;
+use rutil::safe_string::SafeString;
+use rutil::safe_vec::SafeVec;
 use std::env;
 use std::fs::File;
 use std::io::Read;
