+++ projects/rooster/src/list.rs
@@ -173,7 +173,7 @@ mod test {
     use super::get_list_of_passwords;
     use crate::list::{WITHOUT_NUMBERS, WITH_NUMBERS};
     use crate::password::v2::Password;
-    use rutil::rutil::safe_string::SafeString;
+    use rutil::safe_string::SafeString;
 
     // Creates a list of at least two passwords, and more if specified
     fn get_passwords(mut additional: i32) -> Vec<Password> {
