--- projects/rpassword/src/rpassword.rs
@@ -1,3 +1,0 @@
-mod all;
-
-pub use all::*;
