--- projects/rprompt/src/rprompt.rs
@@ -1,3 +1,0 @@
-mod all;
-
-pub use all::{prompt_reply, prompt_reply_from_bufread, read_reply, read_reply_from_bufread};
