+++ projects/rutil/src/lib.rs
@@ -1,1 +1,7 @@
-pub mod rutil;
+pub mod atty;
+pub mod fix_line_issues;
+pub mod print_tty;
+pub mod safe_string;
+#[cfg(feature = "serde")]
+pub mod safe_string_serde;
+pub mod safe_vec;
