--- projects/rutil/src/rutil.rs
@@ -1,7 +1,0 @@
-pub mod atty;
-pub mod fix_line_issues;
-pub mod print_tty;
-pub mod safe_string;
-#[cfg(feature = "serde")]
-pub mod safe_string_serde;
-pub mod safe_vec;
