+++ projects/rutil/src/safe_string.rs
