+++ projects/rutil/src/safe_string_serde.rs
